+++ demo/scripts/lib.combined.js
@@ -1,2 +1,2 @@
-/* updated : 2013-06-20 */
+/* updated : 2013-07-04 */
 myapp.lib.Foo=function(){},myapp.lib.Bar=function(){},myapp.lib.Baz=function(){};