+++ demo/scripts/lib.js
@@ -1,4 +1,4 @@
-head.require(
+_head_.require(
 	"../modules/lib-foo.js",
 	"../modules/lib-bar.js",
 	"../modules/lib-baz.js"