+++ demo/scripts/main.js
@@ -1,4 +1,4 @@
-head.require(
+_head_.require(
 	"../modules/wdg-foo.js", // comment
 	"../modules/wdg-bar.js",
 	// "../modules/hoge.js",